refactor: merge vex behavior into bombastic server
Still have a few naming collisions to workout, mostly around the
common index search structs.
