+++ Cargo.toml
@@ -14,7 +14,6 @@ members = [
     "bommer/bommer-api",
     "common",
     "vexination/vexination",
-    "vexination/api",
     "vexination/index",
     "vexination/model",
     "vexination/indexer",
@@ -65,7 +64,6 @@ default-members = [
     "bommer/bommer-api",
     "common",
     "vexination/vexination",
-    "vexination/api",
     "vexination/index",
     "vexination/model",
     "vexination/indexer",
