+++ bombastic/api/Cargo.toml
@@ -19,6 +19,9 @@ trustification-auth = { path = "../../auth", features = ["actix", "swagger"] }
 trustification-infrastructure = { path = "../../infrastructure" }
 trustification-storage = { path = "../../storage" }
 trustification-index = { path = "../../index" }
+vexination-model = { path = "../../vexination/model" }
+vexination-index = { path = "../../vexination/index" }
+csaf = "0.5.0"
 clap = { version = "4", features = ["derive"] }
 anyhow = "1"
 futures = "0.3"
