+++ integration-tests/src/lib.rs
@@ -3,7 +3,6 @@ mod config;
 mod provider;
 mod spog;
 mod ui;
-mod vex;
 
 pub mod runner;
 
@@ -11,7 +10,6 @@ pub use bom::*;
 pub use provider::*;
 pub use spog::*;
 pub use ui::*;
-pub use vex::*;
 
 use core::future::Future;
 use reqwest::{StatusCode, Url};
