--- vexination/api/Cargo.toml
@@ -1,36 +1,0 @@
-[package]
-name = "vexination-api"
-version = "0.1.0"
-edition = "2021"
-license = "Apache-2.0"
-description = "Vexination REST API for fetching, publishing and searching VEX data"
-
-[dependencies]
-actix-web = "4"
-bytesize = "1"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0.68"
-tokio = { version = "1.0", features = ["full"] }
-log = "0.4"
-trustification-api = { path = "../../api" }
-trustification-auth = { path = "../../auth", features = ["actix", "swagger"] }
-trustification-infrastructure = { path = "../../infrastructure" }
-trustification-storage = { path = "../../storage" }
-trustification-index = { path = "../../index" }
-vexination-index = { path = "../index" }
-vexination-model = { path = "../model" }
-derive_more = "0.99"
-thiserror = "1"
-clap = { version = "4", features = ["derive"] }
-anyhow = "1"
-zstd = "0.12"
-sha2 = "0.10"
-hex = "0.4.3"
-packageurl = "0.3"
-rand = "0.8"
-csaf = "0.5.0"
-prometheus = "0.13.3"
-actix-web-httpauth = "0.8.0"
-
-utoipa = { version = "4", features = ["actix_extras"] }
-utoipa-swagger-ui = { version = "4", features = ["actix-web"] }
