+++ vexination/vexination/src/lib.rs
@@ -2,7 +2,6 @@ use std::process::ExitCode;
 
 #[derive(clap::Subcommand, Debug)]
 pub enum Command {
-    Api(vexination_api::Run),
     Walker(vexination_walker::Run),
     Indexer(vexination_indexer::Run),
 }
@@ -10,7 +9,6 @@ pub enum Command {
 impl Command {
     pub async fn run(self) -> anyhow::Result<ExitCode> {
         match self {
-            Self::Api(run) => run.run(None).await,
             Self::Walker(run) => run.run().await,
             Self::Indexer(run) => run.run().await,
         }
